Character can take damage, HP bar updates
All acceptance tests reached:
- HP decreases when character is reached,
- Damaged animation plays on collision,
- HP bar drops respective to enemy damage.

+++ Mythic Chronicles Sentinel/Assets/Scripts/HealthBar.cs
@@ -7,12 +7,12 @@ public class HealthBar : MonoBehaviour
 {
     public Slider healthSlider;
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         healthSlider.value = health;
     }
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         healthSlider.maxValue = health;
         healthSlider.value = health;
